+++ packages/core/src/preview.ts
@@ -1,7 +1,12 @@
 import { yellow, green, bold } from 'chalk'
-import { CREATE3_PROXY_INITCODE } from '@sphinx-labs/contracts'
+import { CREATE3_PROXY_INITCODE, Operation } from '@sphinx-labs/contracts'
 
-import { DecodedAction, NetworkConfig } from './config/types'
+import {
+  ActionInput,
+  ActionInputType,
+  DecodedAction,
+  NetworkConfig,
+} from './config/types'
 import {
   arraysEqual,
   getNetworkTag,
@@ -9,15 +14,22 @@ import {
   prettyRawFunctionCall,
 } from './utils'
 import { fetchNameForNetwork } from './networks'
+import { InvariantError } from './errors'
 
 type SystemDeploymentElement = {
   type: 'SystemDeployment'
 }
 
+type FundingSafeElement = {
+  type: 'FundingSafe'
+  value: string
+}
+
 type PreviewElement =
   | DecodedAction
-  | { to: string; data: string }
+  | { to: string; data: string; value: string }
   | SystemDeploymentElement
+  | FundingSafeElement
 
 /**
  * @property unlabeledAddresses A set of unlabeled addresses. The preview will warn the user that
@@ -33,6 +45,8 @@ export type SphinxPreview = {
     networkTags: Array<string>
     executing: Array<PreviewElement>
     skipping: Array<PreviewElement>
+    chainId: string
+    safeAddress: string
   }>
   unlabeledAddresses: Set<string>
 }
@@ -57,7 +71,13 @@ export const getPreviewString = (
     `Already executed.`
   )}`
 
-  for (const { networkTags, executing, skipping } of preview.networks) {
+  for (const {
+    networkTags,
+    executing,
+    skipping,
+    chainId,
+    safeAddress,
+  } of preview.networks) {
     // Get the preview string for the networks.
     const networkTagsArray: Array<string> = []
     if (networkTags.length === 1) {
@@ -79,12 +99,15 @@ export const getPreviewString = (
         const element = executing[i]
 
         if (isDecodedAction(element)) {
-          const { referenceName, functionName, variables, address } = element
+          const { referenceName, functionName, variables, address, value } =
+            element
           const actionStr = prettyFunctionCall(
             referenceName,
             address,
             functionName,
             variables,
+            chainId,
+            value,
             5,
             3
           )
@@ -92,6 +115,18 @@ export const getPreviewString = (
           executingArray.push(green(`${i + 1}. ${actionStr}`))
         } else if (isSystemDeploymentElement(element)) {
           executingArray.push(green(`${i + 1}. Sphinx & Gnosis Safe Contracts`))
+        } else if (isFundingSafeElement(element)) {
+          const actionStr = prettyFunctionCall(
+            'GnosisSafe',
+            safeAddress,
+            'call',
+            {},
+            chainId,
+            element.value,
+            5,
+            3
+          )
+          executingArray.push(green(`${i + 1}. ${actionStr}`))
         } else {
           const { to, data } = element
           const actionStr = prettyRawFunctionCall(to, data)
@@ -115,12 +150,18 @@ export const getPreviewString = (
             element.address,
             element.functionName,
             element.variables,
+            chainId,
+            element.value,
             5,
             3
           )
         } else if (isSystemDeploymentElement(element)) {
           throw new Error(
             `Skipped preview elements contain the Sphinx system contracts. Should never happen.`
+          )
+        } else if (isFundingSafeElement(element)) {
+          throw new InvariantError(
+            `Skipped preview elements contain a Safe funding element`
           )
         } else {
           functionCallStr = prettyRawFunctionCall(element.to, element.data)
@@ -157,6 +198,43 @@ export const getPreviewString = (
   return previewString
 }
 
+const assertIsFundingCheckAction = (
+  fundingCheck: ActionInput | undefined,
+  fundingRequest: {
+    fundsRequested: string
+    startingBalance: string
+  },
+  safeAddress: string
+) => {
+  const expectedCheckValue = (
+    BigInt(fundingRequest.fundsRequested) +
+    BigInt(fundingRequest.startingBalance)
+  ).toString()
+
+  if (
+    !fundingCheck ||
+    fundingCheck.txData !== '0x' ||
+    fundingCheck.value !== expectedCheckValue ||
+    fundingCheck.to !== safeAddress ||
+    fundingCheck.actionType !== ActionInputType.CALL ||
+    fundingCheck.requireSuccess !== true ||
+    fundingCheck.operation !== Operation.Call ||
+    fundingCheck.contracts.length !== 0 ||
+    fundingCheck.index !== '1' ||
+    fundingCheck.decodedAction.address !== '' ||
+    fundingCheck.decodedAction.functionName !== 'call' ||
+    fundingCheck.decodedAction.referenceName !== safeAddress ||
+    fundingCheck.decodedAction.value !== expectedCheckValue ||
+    !Array.isArray(fundingCheck.decodedAction.variables) ||
+    fundingCheck.decodedAction.variables?.length !== 1 ||
+    fundingCheck.decodedAction.variables[0] !== '0x'
+  ) {
+    throw new InvariantError(
+      'Expected to find Gnosis Safe funding checking action, but did not'
+    )
+  }
+}
+
 export const getPreview = (
   networkConfigs: Array<NetworkConfig>
 ): SphinxPreview => {
@@ -165,6 +243,8 @@ export const getPreview = (
       executing: Array<PreviewElement>
       skipping: Array<PreviewElement>
       unlabeledAddresses: Array<string>
+      chainId: string
+      safeAddress: string
     }
   } = {}
 
@@ -211,6 +291,7 @@ export const getPreview = (
           functionName: 'deploy',
           variables: {},
           address: networkConfig.safeAddress,
+          value: '0',
         })
       }
       if (!initialState.isModuleDeployed) {
@@ -219,16 +300,49 @@ export const getPreview = (
           functionName: 'deploy',
           variables: {},
           address: networkConfig.moduleAddress,
+          value: '0',
         })
       }
 
-      for (const action of actionInputs) {
-        const { decodedAction } = action
-        executing.push(decodedAction)
+      for (let i = 0; i < actionInputs.length; i++) {
+        /**
+         * We do not display the Safe balance check action in the preview because we don't want to confuse
+         * the user by showing an action that they don't understand without context on why it's there.
+         *
+         * If we find that users are having deployments fail due to this check, then we will reconsider if
+         * we should provide more detail on this specific check.
+         */
+        if (
+          i === 0 &&
+          networkConfig.safeFundingRequest &&
+          BigInt(networkConfig.safeFundingRequest.fundsRequested) > BigInt(0)
+        ) {
+          const [fundingCheck] = actionInputs
+          assertIsFundingCheckAction(
+            fundingCheck,
+            networkConfig.safeFundingRequest,
+            networkConfig.safeAddress
+          )
+
+          // Instead we use a special preview element to represent the Safe funding request
+          executing.push({
+            type: 'FundingSafe',
+            value: networkConfig.safeFundingRequest.fundsRequested,
+          })
+        } else {
+          const { decodedAction } = actionInputs[i]
+          executing.push(decodedAction)
+        }
       }
     }
 
-    networks[networkTag] = { executing, skipping, unlabeledAddresses }
+    networks[networkTag] = {
+      executing,
+      skipping,
+      unlabeledAddresses,
+      chainId: networkConfig.chainId,
+      safeAddress: networkConfig.safeAddress,
+    }
   }
 
   // Next, we group networks that have the same executing and skipping arrays.
@@ -238,7 +352,7 @@ export const getPreview = (
   }
   for (const [
     networkTag,
-    { executing, skipping, unlabeledAddresses },
+    { executing, skipping, unlabeledAddresses, chainId, safeAddress },
   ] of Object.entries(networks)) {
     const existingNetwork = preview.networks.find(
       (e) =>
@@ -256,6 +370,8 @@ export const getPreview = (
         networkTags: [networkTag],
         executing,
         skipping,
+        chainId,
+        safeAddress,
       })
     }
   }
@@ -267,4 +383,10 @@ const isSystemDeploymentElement = (
   element: PreviewElement
 ): element is SystemDeploymentElement => {
   return (element as SystemDeploymentElement).type === 'SystemDeployment'
+}
+
+const isFundingSafeElement = (
+  element: PreviewElement
+): element is FundingSafeElement => {
+  return (element as FundingSafeElement).type === 'FundingSafe'
 }